requirements.txt
exceptiongroup==1.2.2
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.3.2
PyYAML==6.0.1
tomli==2.0.1
